+++ features/transaction/src/main/AndroidManifest.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="me.manulorenzo.moneytransactions.transactions">
+    package="me.manulorenzo.moneytransactions.transaction">
 
     <application>
-        <activity
-            android:name=".TransactionActivity"
-            android:theme="@style/AppTheme.NoActionBar">
+        <activity android:name="me.manulorenzo.moneytransactions.transaction.TransactionActivity">
             <intent-filter>
                 <action android:name="me.manulorenzo.moneytransactions.transaction.open" />
                 <category android:name="android.intent.category.DEFAULT" />
