+++ features/transaction/src/main/res/layout/activity_transaction.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/transactionFragment"
-    class="me.manulorenzo.moneytransactions.transactions.TransactionFragment"
+    class="me.manulorenzo.moneytransactions.transaction.TransactionFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
