--- features/transactions/src/main/res/values/strings.xml
@@ -1,9 +1,0 @@
-<resources>
-    <string name="current_account_number" translatable="false">NL42ABNA7285928462</string>
-    <string name="balance_before_placeholder">Before: %1$s</string>
-    <string name="balance_after_placeholder">After: %1$s</string>
-    <string name="balance_amount_placeholder">%1$s</string>
-    <string name="current_account_placeholder">From: %1$s</string>
-    <string name="destination_account_placeholder">To: %1$s</string>
-    <string name="current_balance_placeholder">%1$s</string>
-</resources>
