+++ libraries/data-transaction/src/main/res/values/strings.xml
