--- libraries/data-transactions/.gitignore
@@ -1,1 +1,0 @@
-/build
